+++ src/test/java/assets/FakerModelAsync.java
@@ -42,16 +42,16 @@ public class FakerModelAsync implements ElsaModel {
     @Id
     private String id;
 
-    @Field(type = FieldType.text, analyzer = "standard")
+    @Field(type = FieldType.Text, analyzer = "standard")
     private String name;
 
-    @Field(type = FieldType.text)
+    @Field(type = FieldType.Text)
     private List<String> favoritePokemons;
 
     @Field(type = FieldType.Integer)
     private int age;
 
-    @Field(type = FieldType.text)
+    @Field(type = FieldType.Text)
     private String biography;