+++ src/test/java/assets/TestModelWithAddedMappings.java
@@ -34,18 +34,18 @@ public class TestModelWithAddedMappings implements ElsaModel {
 
     private String id;
 
-    @Field(type = FieldType.text)
+    @Field(type = FieldType.Text)
     private String stringField;
 
     @Field(type = FieldType.Integer)
     private int integerField;
 
-    @Field(type = FieldType.text)
+    @Field(type = FieldType.Text)
     private List<String> arrayField;
 
     private transient String transientField;
 
-    @Field(type = FieldType.text, analyzer = "standard")
+    @Field(type = FieldType.Text, analyzer = "standard")
     private String addedField;
 
     @Field(type = FieldType.Nested)